+++ test/unit/vagrant/util/line_endings_helper_test.rb
@@ -1,0 +1,16 @@
+require File.expand_path("../../../base", __FILE__)
+
+require "vagrant/util/line_ending_helpers"
+
+describe Vagrant::Util::LineEndingHelpers do
+  let(:klass) do
+    Class.new do
+      extend Vagrant::Util::LineEndingHelpers
+    end
+  end
+
+  it "should convert DOS to unix-style line endings" do
+    klass.dos_to_unix("foo\r\nbar\r\n").should == "foo\nbar\n"
+  end
+end
+